api/app/requirements.txt
asgiref==3.4.1
attrs==21.2.0
bidict==0.21.3
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
fastapi==0.68.1
fastapi-socketio==0.0.8
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
idna==3.2
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
python-engineio==4.2.1
python-socketio==5.4.0
PyYAML==5.4.1
requests==2.26.0
SQLAlchemy==1.4.25
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
